requirements.txt
python-telegram-bot==21.0
groq>=1.0.0
python-dotenv==1.0.1